Issue #94: move annotation reader to common utils
